apply review comments (rename implementation classes, employ Logger to manage error)
