+++ intellijPlugin/src/main/java/com/virtuslab/gitmachete/graph/repositorygraph/data/RepositoryGraphFactory.java
@@ -1,9 +1,9 @@
 package com.virtuslab.gitmachete.graph.repositorygraph.data;
 
 import com.virtuslab.gitmachete.gitmacheteapi.IGitMacheteRepository;
+import com.virtuslab.gitmachete.graph.repositorygraph.BaseRepositoryGraph;
 import com.virtuslab.gitmachete.graph.repositorygraph.RepositoryGraph;
-import com.virtuslab.gitmachete.graph.repositorygraph.RepositoryGraphImpl;
-import com.virtuslab.gitmachete.graph.repositorygraph.RepositoryGraphWithCommitsImpl;
+import com.virtuslab.gitmachete.graph.repositorygraph.RepositoryGraphWithCommits;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import lombok.Getter;
@@ -11,18 +11,18 @@ import lombok.Getter;
 public class RepositoryGraphFactory {
 
   @Getter
-  public static final RepositoryGraph nullRepositoryGraph =
-      new RepositoryGraphImpl(NullRepository.getInstance());
+  public static final BaseRepositoryGraph nullRepositoryGraph =
+      new RepositoryGraph(NullRepository.getInstance());
 
   @Nonnull
-  public static RepositoryGraph getRepositoryGraph(
+  public static BaseRepositoryGraph getRepositoryGraph(
       @Nullable IGitMacheteRepository repository, boolean isListingCommits) {
     if (repository == null) {
       return nullRepositoryGraph;
     } else if (isListingCommits) {
-      return new RepositoryGraphWithCommitsImpl(repository);
+      return new RepositoryGraphWithCommits(repository);
     } else {
-      return new RepositoryGraphImpl(repository);
+      return new RepositoryGraph(repository);
     }
   }
 }
